Don't hide vanished users, but give the invisibily effect instead
This is a temporary fix to 1.19.3 clients disconnecting with `Chat message validation failure`. The cause is the client not being able to find the player who sent the message in the player list. While it is likely possible to not send other clients their player entity while keeping them in the player list, I am not sure if Essentials supports such behavior.